Shifted parsing-related tests to its own file
