+++ http/router_test.go
@@ -28,75 +28,6 @@ var _ = Describe("RequestLineRouter", func() {
 			matchAllRoute = &mock.Route{RouteReturns: &mock.Request{}}
 		})
 
-		Describe("it returns 400 Bad Request", func() {
-			BeforeEach(func() {
-				router = http.NewRouter()
-				router.AddRoute(matchAllRoute)
-			})
-
-			It("for a completely blank request", func() {
-				request, err = router.RouteRequest(makeReader(""))
-				Expect(err).To(beABadRequestResponse("line in request header not ending in CRLF"))
-			})
-
-			It("for any line missing CR", func() {
-				request, err = router.RouteRequest(makeReader("GET / HTTP/1.1\r\n\n"))
-				Expect(err).To(beABadRequestResponse("line in request header not ending in CRLF"))
-			})
-
-			It("for any line missing LF", func() {
-				request, err = router.RouteRequest(makeReader("GET / HTTP/1.1\r"))
-				Expect(err).To(beABadRequestResponse("message header line does not end in LF"))
-			})
-
-			It("for a request missing a request-line", func() {
-				request, err = router.RouteRequest(makeReader("\r\n"))
-				Expect(err).To(beABadRequestResponse("incorrectly formatted or missing request-line"))
-			})
-
-			It("for a request missing an ending CRLF", func() {
-				request, err = router.RouteRequest(makeReader("GET / HTTP/1.1\r\n"))
-				Expect(err).To(beABadRequestResponse("line in request header not ending in CRLF"))
-			})
-
-			It("when multiple spaces are separating fields in request-line", func() {
-				request, err = router.RouteRequest(makeReader("GET /  HTTP/1.1\r\n\r\n"))
-				Expect(err).To(beABadRequestResponse("incorrectly formatted or missing request-line"))
-			})
-
-			It("when fields in request-line contain spaces", func() {
-				request, err = router.RouteRequest(makeReader("GET /a\\ b HTTP/1.1\r\n\r\n"))
-				Expect(err).To(beABadRequestResponse("incorrectly formatted or missing request-line"))
-			})
-
-			It("when the request starts with whitespace", func() {
-				request, err = router.RouteRequest(makeReader(" GET / HTTP/1.1\r\n\r\n"))
-				Expect(err).To(beABadRequestResponse("incorrectly formatted or missing request-line"))
-			})
-		})
-
-		Context("given a well-formed request", func() {
-			var (
-				reader *bufio.Reader
-			)
-
-			BeforeEach(func() {
-				buffer := bytes.NewBufferString("GET /foo HTTP/1.1\r\nAccept: */*\r\n\r\n")
-				reader = bufio.NewReader(buffer)
-				router = http.NewRouter()
-				router.AddRoute(matchAllRoute)
-				request, err = router.RouteRequest(reader)
-			})
-
-			It("returns no error", func() {
-				Expect(err).To(BeNil())
-			})
-
-			It("reads the entire request until reaching a line with only CRLF", func() {
-				Expect(reader.Buffered()).To(Equal(0))
-			})
-		})
-
 		Context("given a well-formed request not matched by any Route", func() {
 			It("returns a NotImplemented response", func() {
 				router = http.NewRouter()
@@ -126,23 +57,6 @@ var _ = Describe("RequestLineRouter", func() {
 			It("returns the request from the first matching Route", func() {
 				Expect(request).To(Equal(matchingRoute.RouteReturns))
 				Expect(err).NotTo(HaveOccurred())
-			})
-		})
-
-		Context("given a well-formed request with query parameters", func() {
-			BeforeEach(func() {
-				router = http.NewRouter()
-				router.AddRoute(matchAllRoute)
-
-				request, err = router.RouteRequest(makeReader("GET /foo?one=1 HTTP/1.1\r\n\r\n"))
-				Expect(err).NotTo(HaveOccurred())
-			})
-
-			XIt("removes the query string from the target", func() {
-				matchAllRoute.ShouldHaveReceived("GET", "/foo")
-			})
-			XIt("passes decoded parameters to the routes", func() {
-				matchAllRoute.ShouldHaveReceivedParameters(map[string]string{"one": "1"})
 			})
 		})
 	})
